lib: views: Fix drop down height mismatch
The drop down menu used to select a model has an inconsistent height on
mobile vs desktop. A height of 40 works correct on desktop, and 50 on
mobile, to match the "Search for inspiration" text field. Update the
height to have a value that matches the search bar. If the "height"
field is removed, it will also get the correct height, but when the
models are still loading, the drop down will have a much smaller height,
which looks bad.